for #660: move asserts to suitable packages
